Fixed text coverage oversight
